+++ NEWS.md
@@ -1,3 +1,8 @@
+Version 0.1.3.999:  November 18, 2019
+-------------------------------------------------------------------------------
++ Adapted `deletionsByVpooled` to infer single chromosome deletion from light chain repertoire.
++ Added html5 interactive output for `deletionHeatmap` visualization.
+
 Version 0.1.3:  September 8, 2019
 -------------------------------------------------------------------------------
 + Optimized `hapHeatmap` visualization, r base graph.
@@ -9,7 +14,7 @@ Version 0.1.2:  July 25, 2019
 Version 0.1.1:  May 8, 2019
 -------------------------------------------------------------------------------
 + Improved `createFullHaplotype` run-time
-+ Fixed a bug werein `createFullHaplotype` miss-handeled novel allele labels
++ Fixed a bug wherein `createFullHaplotype` miss-handled novel allele labels
 
 Version 0.1.0:  March 4, 2019
 -------------------------------------------------------------------------------